+++ R/utils.R
@@ -6,6 +6,23 @@ inverse_map <- function(x, key) {
   names(key)[match(x, key)]
 }
 
+insert_cols <- function(data, pid_cols, pos, replace) {
+  if (replace) {
+    replace_cols(data, pos, pid_cols)
+  } else {
+    add_cols(data, pid_cols, pos)
+  }
+}
+
+replace_cols <- function(data, pos, cols) {
+  stopifnot(length(pos) == length(cols))
+
+  data[pos] <- cols
+  names(data)[pos] <- names(cols)
+
+  data
+}
+
 #' Add new columns to a data frame
 #'
 #' Extends `tibble::add_column` to insert columns to multiple positions.
@@ -14,8 +31,10 @@ inverse_map <- function(x, key) {
 #' @param cols named list of columns to add to `data`
 #' @param pos integer vector of positions to insert columns after
 #' @noRd
-add_cols <- function(data, cols, pos) {
+add_cols <- function(data, cols, pos = ncol(data)) {
   if (length(pos) == 1) {
+    # faster to avoid the loop since `add_column()` already
+    # handles multiple insertion to single position
     data <- tibble::add_column(data, !!!cols, .after = pos)
     return(data)
   }